Never reuse style-sharing ids across or within style passes
Style sharing keys a candidate on its parent's share_id, and that id
was handed out from a counter reset to 0 at the start of every
ns_css_compute, and only to styles entered into the share table. Two
things followed:

- In an incremental pass, parents reused from the previous pass keep
  their old ids while freshly computed parents get new ids from 1 up,
  so an untouched parent and a restyled parent elsewhere could carry
  the same id; a child of one then took the shared style of a child of
  the other and inherited the wrong parent's values. With the rules
  "* + *", ".d", "div.b.c:has(> .a) *.a.b", "div" and "span.a.c",
  removing a class from one <em> left four children of an untouched
  span.q.c without the --r4 their parent span.a.c chain gives them,
  until the next full restyle. The :has() and sibling rules only made
  the tree restyle enough elements to hit a collision.
- Styles never entered into the table (those using attr(), or a
  display: contents element turned into display: none) kept id 0, so
  all their children keyed on the same parent even on a full restyle:
  two .p parents with color: attr(data-c type(<color>)) gave the second
  parent's child the first one's red instead of its own blue.

Make the counter a 64-bit value that is never reset, give every freshly
computed style its own id whether or not it is shared, and give a
shared clone that display_contents_to_none alters a new one. Equal ids
now always mean the same computed style.

A new restyle fuzzer (70 random rules with :has(> x), :has(x),
:has(+ x), :has(~ x), :has(> * > x), attribute and id selectors,
:is/:not, nth-child(of S) and two @scope rules; 50 random class, id and
data-k attribute toggles, each followed by a forced relayout, comparing
the computed values of every element after every step with
NS_NO_INCR_RESTYLE=1): before, 7 of 60 seeds diverged; after, 120/120
seeds match, and 60/60 with element insertions and removals added.
The earlier dyncheck seeds that diverged now match. No measurable perf
change (style pass 210-224 ms on big.html, toggle page unchanged).
Layout snapshots are identical and WPT is unchanged apart from the
known-flaky svg-embedded-sizing tests.

+++ src/css.h
@@ -1089,7 +1089,7 @@ typedef struct ns_style {
     struct ns_style *marker;
     struct ns_style *backdrop;
     struct ns_style *file_selector_button;
-    guint share_id;
+    guint64 share_id;
     int   ref;
     guint32 currentcolor_bits;
     struct ns_var_map *vars;
